change cjs to mjs and small changes
